requirements.txt
django
django-crispy-forms
crispy-bootstrap4
whitenoise
